refactor: apply marketplace component on incorporations checkout
